ml-module/requirements.txt
ultralytics>=8.1.0
numpy>=1.26.0
scipy>=1.11.0
Pillow>=10.0.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
